Revert "Revert "Fixed an issue where writing configuration on stdin should not end with \n as it gives error. It used to work previously, but not with latest versions.""
This reverts commit 685d9026498d0f9ece773b1fa641cb1270ecf0a0.

The reasoning for the previous revert was:

  The initial problem was likely not the lack of \n but a combination
  of not waiting long enough for the child process to be ready, and
  closing the stdin pipe too early.

However that's not correct. The libreswan's configuration parser is messed up.
It tries to seek a pipe to see what's at the end, fails, does not notice the
failure and inserts an extra EOL which the grammar does not allow.

Coincidentally the seek itself is incorrect too and does not do its job for
non-pipe inputs, causing the seek to the beginning and a redundant file parse.
In that case it's rather benign though.

The upstream issue:
https://github.com/libreswan/libreswan/pull/20

Conflicts:
	src/nm-openswan-service.c